ENH: move pv_channel to common tool
